+++ cli.js
@@ -1,6 +1,19 @@
 #!/usr/bin/env node
 
+require('console-stamp')(console, {
+  pattern: 'HH:MM:ss',
+  include: ['log', 'info', 'warn', 'error', 'fatal'],
+  //level: 'log', // development
+  level: 'info', // production
+  colors: {
+    stamp: 'yellow',
+    label: 'yellow',
+    metadata: 'green'
+  }
+});
+
 const os = require('os');
+const log = require('console-log-level')({ level: 'debug' })
 var fs = require('fs');
 var toc = require('./index.js');
 var utils = require('./lib/utils');
@@ -8,80 +21,68 @@ var glob = require('glob');
 var path = require('path');
 var args = utils.minimist(process.argv.slice(2), {
   boolean: ['i'],
-  string: ['f'],
   string: ['d'],
-  // default: {
-  //   d: path.resolve(process.cwd()),
-  //   f: path.resolve(process.cwd()) + '/index.md'
-  // },
   alias: {h: 'help'}
 });
 
-
-if (!args.d && !args.f && !args.i || args.h) {
-  console.error([
-    'Usage: adr-log [-i] [-d <directory>] [-f <file>]',
+if (!args.d && !args.i && (args._.length==0) || args.h) {
+  process.stderr.write([
+    'Usage: adr-log [-d <directory>] [-i] <input>',
     '',
-    '  directory:  The directory to be scanned for the *.md files',
-    '              If no <directory> is given, the current working directory will be chosen by default',
+    '  input:  The markdown file to contain the table of contents.',
+    '          If no <input> file is specified, an index.md file containing the log is created in the current directory.',
     '',
-    '  file:      The markdown file to contain the table of contents,',
-    '              If no <file> file is specified, a index.md file containing the log is created in the current directory.',
+    '  -i:     Edit the <input> file directly, injecting the log at <!-- adrlog -->.',
+    '          Using only the -i flag, the tool will scan the current working directory for all *.md files and',
+    '          inject the resulting adr-log into the default index.md file.',
+    '          (Without this flag, the default is to print the log to stdout.)',
     '',
-    '  -i:         Edit the <file> file directly, injecting the log at <!-- adrlog -->',
-    '              Using only the -i flag, the tool will scan the current working directory for all *.md files and inject the resulting adr-log into the default index.md file ',
-    '              (Without this flag, the default is to print the log to stdout.)',
+    '  -d:     Scans the given <directory> for .md files.',
+    '          (Without this flag, the current working directory is chosen as default.)',
     '',
-    '  -d:         Scans the given <directory> for .md files and adds them to the log which gets injected into the <file>.',
-    '              (Without this flag, the current working directory will be chosen as default)',
-    '',
-    '  -f:         Option to specify the <file> in which the adr-log should be injected',
-    '              (Without this flag index.md will be chosen as default.)'
+    '  -h:     Shows how to use this program',
+    ''
   ].join(os.EOL));
   process.exit(1);
 }
 
 if (args.i && args._[0] === '-') {
-  console.error('adr-log: you cannot use -i with "-" (stdin) for input');
+  process.stderr.write('adr-log: you cannot use -i with "-" (stdin) for input');
   process.exit(1);
 }
 
-var input = process.stdin;
+var defaultAdrLogDir = path.resolve(process.cwd());
+var adrLogDir = args.d || defaultAdrLogDir;
 
-var defaultDir = path.resolve(process.cwd());
-var defaultFile = 'index.md';
-var dir = args.d || defaultDir;
-var tocFile = args.f || defaultFile;
+var defaultAdrLogFile = 'index.md';
+var adrLogFile = args._[0] || adrLogDir + '/' + defaultAdrLogFile;
+var adrLogFileName = require('path').parse(adrLogFile).base;
+
+console.log("adr log file:", adrLogFile);
+console.log("adr log dir:", adrLogDir);
 
 var headings = '';
-var filenames = glob.sync('!(' + tocFile.slice(0,-3) + '*).md', {cwd: dir});
+const globPattern = '!(' + adrLogFileName.slice(0,-3) + '*).md';
+console.log("glob pattern:", globPattern);
+var filenames = glob.sync(globPattern, {cwd: adrLogDir});
+console.log("filenames:", filenames);
 
+// determine log entries
 for (const filename of filenames) {
+  console.log("add filename:", filename);
   headings += utils.headify(filename + '\n');
 }
 
-if (args.i && fs.existsSync(tocFile)) {
-  input = fs.createReadStream(tocFile);
-
-  input.pipe(utils.concat(function (input) {
-    var newMarkdown = toc.insertAdrToc(input.toString(), headings, dir);
-    fs.writeFileSync(tocFile, newMarkdown);
-  }));
-} else if (args.i) {
-  var tocString = '<!-- adrlog -->' + os.EOL + os.EOL + '<!-- adrlogstop -->' + os.EOL;
-
-  fs.writeFileSync(tocFile, toc.insertAdrToc(tocString, headings, dir));
+var existingLogString;
+if (fs.existsSync(adrLogFile)) {
+  existingLogString = fs.readFileSync(adrLogFile, 'utf8');
 } else {
-  var parsed = toc(headings, dir);
-  output(parsed);
+  existingLogString = '<!-- adrlog -->' + os.EOL + os.EOL + '<!-- adrlogstop -->' + os.EOL;
 }
+var newLogString = toc.insertAdrToc(existingLogString, headings, adrLogDir);
 
-input.on('error', function onErr(err) {
-  console.error(err);
-  process.exit(1);
-});
-
-function output(parsed) {
-  if (args.json) return console.log(JSON.stringify(parsed.json, null, '  '));
-  process.stdout.write(parsed.content);
+if (args.i) {
+  fs.writeFileSync(adrLogFile, newLogString);
+} else {
+  process.stdout.write(newLogString);
 }