+++ CHANGELOG.md
@@ -11,9 +11,13 @@ We refer to [GitHub issues](https://github.com/adr/adr-log/issues) by using `#NU
 ### Added
 - Support for comments after `adrlog` in `<!-- adrlog -->`. [#17](https://github.com/adr/adr-log/issues/17)
 
+### Changed
+- Updated CLI syntax to be more conform with [markdown-toc](https://github.com/jonschlinkert/markdown-toc).
+
 ### Fixed
 - The title of the first main heading is used (and not the second line if it is not empty). [#15](https://github.com/adr/adr-log/issues/15)
 - When adrlog appears without closing adrlogstop, adr-log does not stop with an error anymore.
+- When adrlog does not appear, the log is created nevertheless
 
 ## [v1.1.0] - 2017-10-27